refactor(interpreter): relocate resolver code from interpreter to values
Also, make compiler.funtionValue implement the Resolver interface.
